feat: extract jwt generation so storage hosts can also make use of it
